improve abstraction, finish dirs
