--- src/Schema/Directives/MutationExecutorDirective.php
@@ -1,88 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Nuwave\Lighthouse\Schema\Directives;
-
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\Relation;
-use Nuwave\Lighthouse\Execution\Arguments\ArgumentSet;
-use Nuwave\Lighthouse\Execution\Arguments\ResolveNested;
-use Nuwave\Lighthouse\Execution\ResolveInfo;
-use Nuwave\Lighthouse\Execution\TransactionalMutations;
-use Nuwave\Lighthouse\Schema\Values\FieldValue;
-use Nuwave\Lighthouse\Support\Contracts\ArgResolver;
-use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
-use Nuwave\Lighthouse\Support\Utils;
-
-abstract class MutationExecutorDirective extends BaseDirective implements FieldResolver, ArgResolver
-{
-    public function __construct(
-        protected TransactionalMutations $transactionalMutations,
-    ) {}
-
-    public function resolveField(FieldValue $fieldValue): callable
-    {
-        $modelClass = $this->getModelClass();
-
-        return function (mixed $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($modelClass): Model {
-            $model = new $modelClass();
-
-            return $this->transactionalMutations->execute(
-                function () use ($model, $resolveInfo): Model {
-                    $mutated = $this->executeMutation($model, $resolveInfo->argumentSet);
-                    assert($mutated instanceof Model);
-
-                    return $mutated->refresh();
-                },
-                $model->getConnectionName(),
-            );
-        };
-    }
-
-    /**
-     * @param  Model  $model
-     * @param  \Nuwave\Lighthouse\Execution\Arguments\ArgumentSet|array<\Nuwave\Lighthouse\Execution\Arguments\ArgumentSet>  $args
-     *
-     * @return \Illuminate\Database\Eloquent\Model|array<\Illuminate\Database\Eloquent\Model>
-     */
-    public function __invoke($model, $args): mixed
-    {
-        $relationName = $this->directiveArgValue(
-            'relation',
-            // Use the name of the argument if no explicit relation name is given
-            $this->nodeName(),
-        );
-
-        $relation = $model->{$relationName}();
-        assert($relation instanceof Relation);
-
-        // @phpstan-ignore-next-line Relation&Builder mixin not recognized
-        $related = $relation->make();
-        assert($related instanceof Model);
-
-        return $this->executeMutation($related, $args, $relation);
-    }
-
-    /**
-     * @param  \Nuwave\Lighthouse\Execution\Arguments\ArgumentSet|array<\Nuwave\Lighthouse\Execution\Arguments\ArgumentSet>  $args
-     * @param  \Illuminate\Database\Eloquent\Relations\Relation<\Illuminate\Database\Eloquent\Model>|null  $parentRelation
-     *
-     * @return \Illuminate\Database\Eloquent\Model|array<\Illuminate\Database\Eloquent\Model>
-     */
-    protected function executeMutation(Model $model, ArgumentSet|array $args, ?Relation $parentRelation = null): Model|array
-    {
-        $update = new ResolveNested($this->makeExecutionFunction($parentRelation));
-
-        return Utils::mapEach(
-            static fn (ArgumentSet $argumentSet): mixed => $update($model->newInstance(), $argumentSet),
-            $args,
-        );
-    }
-
-    /**
-     * Prepare the execution function for a mutation on a model.
-     *
-     * @param  \Illuminate\Database\Eloquent\Relations\Relation<\Illuminate\Database\Eloquent\Model>|null  $parentRelation
-     */
-    abstract protected function makeExecutionFunction(?Relation $parentRelation = null): callable;
-}
